back/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.23.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.7.1
requests==2.31.0
bcrypt==4.0.1

